requirements/prod.txt
# This file was autogenerated by uv v0.1.2 via the following command:
#    uv pip compile requirements/prod.in -o requirements/prod.txt
aiohttp==3.9.3
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via langchain-core
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via
    #   halo
    #   log-symbols
dataclasses-json==0.6.4
    # via
    #   langchain
    #   langchain-community
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via sqlalchemy
halo==0.0.31
idna==3.6
    # via
    #   anyio
    #   requests
    #   yarl
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
langchain==0.1.7
langchain-community==0.0.20
    # via langchain
langchain-core==0.1.23
    # via
    #   langchain
    #   langchain-community
langsmith==0.0.87
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
log-symbols==0.0.14
    # via halo
markdown-it-py==3.0.0
    # via rich
marshmallow==3.20.2
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
packaging==23.2
    # via
    #   langchain-core
    #   marshmallow
pydantic==2.6.1
    # via
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.16.2
    # via pydantic
pygments==2.17.2
    # via rich
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
requests==2.31.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langsmith
rich==13.7.0
six==1.16.0
    # via halo
sniffio==1.3.0
    # via anyio
spinners==0.0.24
    # via halo
sqlalchemy==2.0.27
    # via
    #   langchain
    #   langchain-community
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
termcolor==2.4.0
    # via halo
typing-extensions==4.9.0
    # via
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.2.0
    # via requests
yarl==1.9.4
    # via aiohttp
